+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * Add `StreamListenerInterface` to hook into response streams
  * [BC BREAK] Change `TokenUsageAggregation::__construct()` from variadic to array
+ * Add `TokenUsageAggregation::add()` method to add more token usages
 
 0.2
 ---